+++ tools/gvcore/files/gvcore.py
@@ -30,7 +30,7 @@ try:
 except ImportError:
     paramiko = None
 
-__version__ = "2.0.0"
+__version__ = "2.0.1"
 
 # ─────────────────────────────────────────────────────────────────────────────
 # XDG Paths